backend/pyproject.toml
[project]
name = "image-framer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.6",
    "fastapi[standard]>=0.116.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
]
